% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similiars.R
\name{case_when_skeleton}
\alias{case_when_skeleton}
\title{case_when_skeleton}
\usage{
case_when_skeleton(.s, .r, var_name = NULL)
}
\arguments{
\item{.s}{a character vector to be replaced.}

\item{.r}{a character vector to replace with.}

\item{var_name}{Optional. the name of the variable that should be replaced. If NULL or not changed the call of '.s' is used.}
}
\value{
console output with a case_when skeleton
}
\description{
Returns a \code{case_when} from \code{dplyr} to replace .s with .r. You can copy the output and put in the script to avoid manual work.
}
\examples{

case_when_skeleton(letters[1:10],
                   LETTERS[1:10],
                   var_name = "my_letters")

}
