% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similiars.R
\name{find_most_similar_string}
\alias{find_most_similar_string}
\title{find_most_similar_string}
\usage{
find_most_similar_string(
  .s,
  .t,
  max_dist = Inf,
  verbose = TRUE,
  ignore_case = TRUE,
  feeling_lucky = FALSE,
  ...
)
}
\arguments{
\item{.s}{a character vector to be matched.}

\item{.t}{a character vector to be matched against.}

\item{max_dist}{the maximum string distance}

\item{verbose}{should warnings be printed in the console.}

\item{ignore_case}{should case be ignored? Default is TRUE.}

\item{feeling_lucky}{if multiple most similar strings are found. Should the first one be returned?}

\item{...}{other arguments passed to utils::adist.}
}
\value{
a character vector
}
\description{
Returns a vector of most similar string in another vector. 
Returns a vector of the same length as input vector '.s'.
}
\details{
Uses the generalized Levenshtein distance. For more information type \code{?utils::adist} in the console.
}
\examples{

find_most_similar_string(c("pple", "peacj"), stringr::fruit)

}
