% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similiars.R
\name{find_string_distance}
\alias{find_string_distance}
\title{find_string_distance}
\usage{
find_string_distance(.s, .t, ignore_case = TRUE, ...)
}
\arguments{
\item{.s}{a character vector to be matched.}

\item{.t}{a character vector to be matchad against.}

\item{ignore_case}{should case be ignored? Default is TRUE.}

\item{...}{other arguments passed to utils::adist.}
}
\value{
a list of dataframes
}
\description{
Returns a list of dataframes arranged by most similar strings.
}
\details{
Uses the generalized Levenshtein distance. For more information type \code{?utils::adist} in the console.
}
\examples{

find_string_distance("pple", stringr::fruit)

}
