% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similiars.R
\docType{data}
\name{superheroes}
\alias{superheroes}
\title{Famous super heroes}
\format{
A character vector with 46 super heroes:
}
\source{
\url{https://www.britannica.com/topic/list-of-superheroes-2024795}
}
\usage{
superheroes
}
\description{
A character vector with 46 super heroes
}
\keyword{datasets}
